backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Limiter==3.3.0
bcrypt==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.1
gunicorn==21.2.0